Using logging middleware to report errors instead of separate middleware
