+++ json/handler.go
@@ -3,8 +3,6 @@ package json
 import (
 	"net/http"
 	"time"
-
-	"github.com/BSick7/go-api/errors"
 )
 
 type HandlerFunc func(res *ResponseWriter, req *Request)
@@ -12,11 +10,7 @@ type HandlerFunc func(res *ResponseWriter, req *Request)
 func Handler(handler HandlerFunc) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		w.Header().Set("Content-Type", "application/json")
-		res := &ResponseWriter{
-			ResponseWriter: w,
-			start:          time.Now(),
-			errContainer:   errors.ContainerFromContext(r.Context()),
-		}
+		res := &ResponseWriter{ResponseWriter: w, start: time.Now()}
 		req := &Request{Request: r}
 		handler(res, req)
 	})
