+++ json/response_writer.go
@@ -5,8 +5,6 @@ import (
 	"fmt"
 	"net/http"
 	"time"
-
-	"github.com/BSick7/go-api/errors"
 )
 
 type HttpError struct {
@@ -20,20 +18,13 @@ func (e HttpError) Error() string {
 
 type ResponseWriter struct {
 	http.ResponseWriter
-	start        time.Time
-	statusCode   int
-	errContainer *errors.Container
+	start      time.Time
+	statusCode int
 }
 
 func (r *ResponseWriter) SendError(statusCode int, err error) {
 	r.statusCode = statusCode
 	r.ResponseWriter.WriteHeader(statusCode)
-	if r.errContainer != nil {
-		r.errContainer.AddError(&HttpError{
-			StatusCode: statusCode,
-			Err:        err,
-		})
-	}
 	encoder := json.NewEncoder(r.ResponseWriter)
 	encoder.Encode(map[string]interface{}{"error": err.Error()})
 }
