+++ jsonapi/handler.go
@@ -4,7 +4,6 @@ import (
 	"net/http"
 	"time"
 
-	"github.com/BSick7/go-api/errors"
 	"github.com/svanharmelen/jsonapi"
 )
 
@@ -13,11 +12,7 @@ type HandlerFunc func(res *ResponseWriter, req *Request)
 func Handler(handler HandlerFunc) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		w.Header().Set("Content-Type", jsonapi.MediaType)
-		res := &ResponseWriter{
-			ResponseWriter: w,
-			start:          time.Now(),
-			errContainer:   errors.ContainerFromContext(r.Context()),
-		}
+		res := &ResponseWriter{ResponseWriter: w, start: time.Now()}
 		req := &Request{Request: r}
 		handler(res, req)
 	})
