+++ logging/endpoint_logger.go
@@ -10,45 +10,6 @@ import (
 	"github.com/gorilla/mux"
 )
 
-type Config struct {
-	// Prefix is prepended to all log statements
-	Prefix string
-
-	// Log100s will determine whether an HTTP 1xx response is logged
-	Log100s bool
-
-	// Log200s will determine whether an HTTP 2xx response is logged
-	Log200s bool
-
-	// Log300s will determine whether an HTTP 3xx response is logged
-	Log300s bool
-
-	// Log400s will determine whether an HTTP 4xx response is logged
-	Log400s bool
-
-	// Log500s will determine whether an HTTP 5xx response is logged
-	Log500s bool
-}
-
-func (c Config) ShouldLog(statusCode int) bool {
-	if statusCode >= 100 && statusCode < 200 {
-		return c.Log100s
-	}
-	if statusCode >= 200 && statusCode < 300 {
-		return c.Log200s
-	}
-	if statusCode >= 300 && statusCode < 400 {
-		return c.Log300s
-	}
-	if statusCode >= 400 && statusCode < 500 {
-		return c.Log400s
-	}
-	if statusCode >= 500 {
-		return c.Log500s
-	}
-	return false
-}
-
 func EndpointLoggerMiddleware(cfg Config) mux.MiddlewareFunc {
 	stdNoTime := log.New(os.Stderr, cfg.Prefix, 0)
 
@@ -57,17 +18,31 @@ func EndpointLoggerMiddleware(cfg Config) mux.MiddlewareFunc {
 			start := time.Now()
 			wrapped := &endpointLoggerWriterWrapper{ResponseWriter: w, statusCode: http.StatusOK}
 			handler.ServeHTTP(wrapped, r)
+			dur := time.Since(start)
+			if cfg.OnRequest != nil {
+				cfg.OnRequest(r, wrapped, dur)
+			}
 			if cfg.ShouldLog(wrapped.statusCode) {
-				stdNoTime.Printf("%s %d %s %s %s", time.Since(start), wrapped.statusCode, r.Method, r.RequestURI, strings.Join(wrapped.capturedData, ""))
+				stdNoTime.Printf("%s %d %s %s %s", dur, wrapped.StatusCode(), r.Method, r.RequestURI, wrapped.Body())
 			}
 		})
 	}
 }
 
+var _ ResponseData = &endpointLoggerWriterWrapper{}
+
 type endpointLoggerWriterWrapper struct {
 	http.ResponseWriter
 	statusCode   int
 	capturedData []string
+}
+
+func (w *endpointLoggerWriterWrapper) StatusCode() int {
+	return w.statusCode
+}
+
+func (w *endpointLoggerWriterWrapper) Body() string {
+	return strings.Join(w.capturedData, "")
 }
 
 func (w *endpointLoggerWriterWrapper) Header() http.Header {
